feat: Add Automated line counting for Resume Info.
Since I'd like to include Line Counts for some of the projects listed on my Resume, I decided to create this workflow that runs on the 1st day of every month, only if there have been commits/pushes since the last successful run of this workflow (changes to the code have been made since the last run). This workflow counts the lines of code, by language used and it is found in the artifact section of the workflow.